Fix int overflow problem

+++ ioctl.go
@@ -5,12 +5,12 @@ import (
 	"unsafe"
 )
 
-func winIoctl(fd int, cmd int, win *winsize) {
-	_, _, _ = syscall.Syscall(syscall.SYS_IOCTL, uintptr(fd), uintptr(cmd), uintptr(unsafe.Pointer(win)))
+func winIoctl(fd int, cmd uintptr, win *winsize) {
+	_, _, _ = syscall.Syscall(syscall.SYS_IOCTL, uintptr(fd), cmd, uintptr(unsafe.Pointer(win)))
 	return
 }
 
-func ttyIoctl(fd int, cmd int, term *termios) {
-	_, _, _ = syscall.Syscall(syscall.SYS_IOCTL, uintptr(fd), uintptr(cmd), uintptr(unsafe.Pointer(term)))
+func ttyIoctl(fd int, cmd uintptr, term *termios) {
+	_, _, _ = syscall.Syscall(syscall.SYS_IOCTL, uintptr(fd), cmd, uintptr(unsafe.Pointer(term)))
 	return
 }
